refactor(TransactionsRoutes): move component to its own directory
